Moved test services under `test_service/`.
