+++ examples/Karl/service_generator.cc
@@ -1,6 +1,6 @@
 #include "../../Blocks/HTTP/api.h"
 #include "../../Bricks/dflags/dflags.h"
-#include "../../Karl/service_generator.h"
+#include "../../Karl/test_service/generator.h"
 
 DEFINE_uint16(port, 42001, "The port to spawn ServiceGenerator on.");
 DEFINE_double(nps, 10.0, "The `numbers per second` to generate.");
